Split Header into HeaderTitle and HeaderRight components
